Migrating shape profiles
- Moving all shape profiles from Geometry_oM to Spatial_oM
- Updating namespace of all ShapeProfiles
- Adding references where needed to Spatial_oM
- Updating namespaces
- Also adding descriptions for most profiles
