+++ Physical_oM/FramingProperties/ConstantFramingProperty.cs
@@ -27,7 +27,7 @@ using System.Text;
 using System.Threading.Tasks;
 
 using BH.oM.Base;
-using BH.oM.Geometry.ShapeProfiles;
+using BH.oM.Spatial.ShapeProfiles;
 using BH.oM.Physical.Materials;
 
 namespace BH.oM.Physical.FramingProperties
