+++ Spatial_oM/ShapeProfiles/AngleProfile.cs
@@ -25,9 +25,12 @@ using BH.oM.Geometry;
 using System.Collections.ObjectModel;
 using System.Collections.Generic;
 using System.Linq;
+using System.ComponentModel;
+using BH.oM.Quantities.Attributes;
 
-namespace BH.oM.Geometry.ShapeProfiles
+namespace BH.oM.Spatial.ShapeProfiles
 {
+    [Description("L-shaped profile")]
     public class AngleProfile : BHoMObject, IProfile, IImmutable
     {
 
@@ -36,22 +39,37 @@ namespace BH.oM.Geometry.ShapeProfiles
         /***************************************************/
         public virtual ShapeType Shape { get; } = ShapeType.Angle;
 
+        [Length]
+        [Description("Full depth of the profile from bottom of flange to top of the web.")]
         public virtual double Height { get; }
 
+        [Length]
+        [Description("Full width of the profile from the furthest side of the web to the end of the flange.")]
         public virtual double Width { get; }
 
+        [Length]
+        [Description("Thickness of the stem of the L-shape.")]
         public virtual double WebThickness { get; }
 
+        [Length]
+        [Description("Thickness of the base of the L-shape.")]
         public virtual double FlangeThickness { get; }
 
+        [Length]
+        [Description("Fillet radius between inner face of the flange and inner face of the web.")]
         public virtual double RootRadius { get; }
 
+        [Length]
+        [Description("Fillet radius at the outer edge of the flange and web. Value need to be smaller or equal than the Flange and Web thicknesses")]
         public virtual double ToeRadius { get; }
 
+        [Description("If true, the section is mirrored about its local z-axis, resulting in a backwards facing L-shape.")]
         public virtual bool MirrorAboutLocalZ { get; }
 
+        [Description("If true, the section is mirrored about its local y-axis, resulting in upside down L-shape.")]
         public virtual bool MirrorAboutLocalY { get; }
 
+        [Description("Edge curves that matches the dimensions in the global XY-plane.")]
         public virtual ReadOnlyCollection<ICurve> Edges { get; }
