+++ Spatial_oM/Spatial_oM.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -51,6 +51,24 @@
     <Compile Include="Layouts\PerimeterLayout.cs" />
     <Compile Include="Layouts\Enums\ReferencePoint.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ShapeProfiles\AngleProfile.cs" />
+    <Compile Include="ShapeProfiles\BoxProfile.cs" />
+    <Compile Include="ShapeProfiles\ChannelProfile.cs" />
+    <Compile Include="ShapeProfiles\CircleProfile.cs" />
+    <Compile Include="ShapeProfiles\Enums\ShapeType.cs" />
+    <Compile Include="ShapeProfiles\FabricatedBoxProfile.cs" />
+    <Compile Include="ShapeProfiles\FabricatedISectionProfile.cs" />
+    <Compile Include="ShapeProfiles\FreeFormProfile.cs" />
+    <Compile Include="ShapeProfiles\GeneralisedFabricatedBoxProfile.cs" />
+    <Compile Include="ShapeProfiles\GeneralisedTSectionProfile.cs" />
+    <Compile Include="ShapeProfiles\IProfile.cs" />
+    <Compile Include="ShapeProfiles\ISectionProfile.cs" />
+    <Compile Include="ShapeProfiles\KiteProfile.cs" />
+    <Compile Include="ShapeProfiles\RectangleProfile.cs" />
+    <Compile Include="ShapeProfiles\TaperedProfile.cs" />
+    <Compile Include="ShapeProfiles\TSectionProfile.cs" />
+    <Compile Include="ShapeProfiles\TubeProfile.cs" />
+    <Compile Include="ShapeProfiles\ZSectionProfile.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\BHoM\BHoM.csproj">
@@ -69,6 +87,9 @@
       <Project>{29e6dcd7-270a-45cc-ac0b-6c024287645e}</Project>
       <Name>Reflection_oM</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Versioning_40.json" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
